Single status bar on top
Use sketchybar as the top status bar, with no bar at the bottom and the
stock menu bar hidden.

This may be reverted later in favor of the simpler earlier approach
where the stock menu bar was on top, and sketchybar was on the bottom.
This version has some hacks to deal with different bar heights required
whether or not the display is notched.

This also adds the stats package to the bar and uses sketchybar aliases
to display stock menu bar items. These may be replaced by custom shell
scripts later.